Create AbstractCryptor that is base class of all cryptors
