+++ src/encryptdecrypt/Main.java
@@ -1,5 +1,7 @@
 package encryptdecrypt;
 
+import encryptdecrypt.cryptors.AbstractCryptor;
+import encryptdecrypt.cryptors.ConcreteCryptor;
 import encryptdecrypt.parsers.CliParser;
 import encryptdecrypt.strategy.ShiftAlgorithm;
 import encryptdecrypt.strategy.UnicodeAlgorithm;
@@ -9,7 +11,7 @@ public class Main {
     public static void main(String[] args) {
         CliParser parser = new CliParser(args);
 
-        EncryptorDecryptor encryptorDecryptor = new EncryptorDecryptor(
+        AbstractCryptor cryptor = new ConcreteCryptor(
                 parser.optionValueOrDefault("-alg", "shift")
                         .equals("shift") ? new ShiftAlgorithm() : new UnicodeAlgorithm(),
 
@@ -20,6 +22,6 @@ public class Main {
                 Integer.parseInt(parser.optionValueOrDefault("-key", "0"))
         );
 
-        encryptorDecryptor.start();
+        cryptor.start();
     }
 }
